requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
freeflow-llm==0.1.1
python-dotenv==1.0.1

